requirements.txt
# Dépendances principales
Pillow>=10.0.0
